Split UloopPausePointRegistry co-located types into class-per-file
UloopPausePointRegistry.cs mixed the registry facade with 8 other
top-level types (snapshot, entry, pause controller, DTOs, status
constants) in one 525-line file. Move each type to its own file,
byte-equal, keeping the registry's mutable static state at its
original path so no test tracking needs updating.
