+++ Packages/src/Runtime/PausePoints/IUloopPausePointPauseController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 01582b4ec2824e59a7cf433fb3c29ec8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
