+++ Packages/src/Runtime/PausePoints/UloopPausePointClearAllResult.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 966652563c754499ba5e308f5124aed1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
