+++ Packages/src/Runtime/PausePoints/UloopPausePointEditorStateCapturedAt.cs
@@ -1,0 +1,14 @@
+#if UNITY_EDITOR
+namespace io.github.hatayama.UnityCliLoop.Runtime
+{
+    /// <summary>
+    /// Names when a pause point editor-state snapshot was captured.
+    /// </summary>
+    internal static class UloopPausePointEditorStateCapturedAt
+    {
+        public const string Current = "Current";
+        public const string PausePointHit = "PausePointHit";
+        public const string ClearAll = "ClearAll";
+    }
+}
+#endif
