+++ Packages/src/Runtime/PausePoints/UloopPausePointEditorStateCapturedAt.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3a72ea04c8274ba5a1a9d6c1ac982051
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
