+++ Packages/src/Runtime/PausePoints/UloopPausePointEditorStateSnapshot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9e9962bcd864491c9f386ba45bd53ba0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
