+++ Packages/src/Runtime/PausePoints/UloopPausePointEntry.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ee44c84e35fb49bfa50d1d3844adfefd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
