+++ Packages/src/Runtime/PausePoints/UloopPausePointStatus.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 430639d35bfd48e79b18337984adcfca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
