+++ Packages/src/Runtime/PausePoints/UnityEditorPausePointPauseController.cs
@@ -1,0 +1,20 @@
+#if UNITY_EDITOR
+using UnityEditor;
+
+namespace io.github.hatayama.UnityCliLoop.Runtime
+{
+    /// <summary>
+    /// Adapts pause point hits to UnityEditor.EditorApplication state.
+    /// </summary>
+    internal sealed class UnityEditorPausePointPauseController : IUloopPausePointPauseController
+    {
+        public bool IsPlaying => EditorApplication.isPlaying;
+        public bool IsPaused => EditorApplication.isPaused;
+
+        public void Pause()
+        {
+            EditorApplication.isPaused = true;
+        }
+    }
+}
+#endif
