+++ Packages/src/Runtime/PausePoints/UnityEditorPausePointPauseController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6028284fed914698bba0b40c49ae48d9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData:
+  assetBundleName:
+  assetBundleVariant:
